+++ src/Event/AbstractTestEvent.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Event;
 
-use App\Entity\Test;
 use Symfony\Contracts\EventDispatcher\Event;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
 
 abstract class AbstractTestEvent extends Event
 {
