+++ src/Model/RunnerTest/TestConfigurationProxy.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Model\RunnerTest;
 
-use App\Entity\TestConfiguration as TestConfigurationEntity;
 use webignition\BasilRunnerDocuments\TestConfiguration;
+use webignition\BasilWorker\PersistenceBundle\Entity\TestConfiguration as TestConfigurationEntity;
 
 class TestConfigurationProxy extends TestConfiguration
 {
