+++ src/Services/TestFactory.php
@@ -4,27 +4,20 @@ declare(strict_types=1);
 
 namespace App\Services;
 
-use App\Entity\Test;
-use App\Entity\TestConfiguration;
 use App\Event\SourceCompile\SourceCompileSuccessEvent;
-use App\Repository\TestRepository;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use webignition\BasilCompilerModels\TestManifest;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Entity\TestConfiguration;
+use webignition\BasilWorker\PersistenceBundle\Services\Factory\TestFactory as BundleTestFactory;
 
 class TestFactory implements EventSubscriberInterface
 {
-    private TestStore $testStore;
-    private TestRepository $testRepository;
-    private TestConfigurationStore $testConfigurationStore;
+    private BundleTestFactory $bundleTestFactory;
 
-    public function __construct(
-        TestStore $testStore,
-        TestRepository $testRepository,
-        TestConfigurationStore $testConfigurationStore
-    ) {
-        $this->testStore = $testStore;
-        $this->testRepository = $testRepository;
-        $this->testConfigurationStore = $testConfigurationStore;
+    public function __construct(BundleTestFactory $bundleTestFactory)
+    {
+        $this->bundleTestFactory = $bundleTestFactory;
     }
 
     public static function getSubscribedEvents()
@@ -70,7 +63,7 @@ class TestFactory implements EventSubscriberInterface
     {
         $manifestConfiguration = $manifest->getConfiguration();
 
-        return $this->create(
+        return $this->bundleTestFactory->create(
             TestConfiguration::create(
                 $manifestConfiguration->getBrowser(),
                 $manifestConfiguration->getUrl()
@@ -79,32 +72,5 @@ class TestFactory implements EventSubscriberInterface
             $manifest->getTarget(),
             $manifest->getStepCount()
         );
-    }
-
-    protected function create(
-        TestConfiguration $configuration,
-        string $source,
-        string $target,
-        int $stepCount
-    ): Test {
-        $position = $this->findNextPosition();
-        $configuration = $this->testConfigurationStore->findByConfiguration($configuration);
-        $test = Test::create($configuration, $source, $target, $stepCount, $position);
-
-        return $this->testStore->store($test);
-    }
-
-    private function findNextPosition(): int
-    {
-        $maxPosition = $this->findMaxPosition();
-
-        return null === $maxPosition
-            ? 1
-            : $maxPosition + 1;
-    }
-
-    private function findMaxPosition(): ?int
-    {
-        return $this->testRepository->findMaxPosition();
     }
 }
