+++ src/Services/TestSerializer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Services;
 
-use App\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
 
 class TestSerializer
 {
