+++ tests/Integration/Asynchronous/EndToEnd/CreateAddSourcesCompileExecuteTest.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Integration\Asynchronous\EndToEnd;
 
-use App\Entity\Test;
 use App\Model\BackoffStrategy\ExponentialBackoffStrategy;
-use App\Repository\TestRepository;
 use App\Services\ApplicationState;
 use App\Services\CompilationState;
 use App\Services\ExecutionState;
@@ -20,6 +18,8 @@ use App\Tests\Services\InvokableFactory\JobSetup;
 use App\Tests\Services\InvokableFactory\TestGetterFactory;
 use Psr\Http\Message\RequestInterface;
 use webignition\BasilWorker\PersistenceBundle\Entity\Callback\CallbackInterface;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Services\Repository\TestRepository;
 use webignition\SymfonyTestServiceInjectorTrait\TestClassServicePropertyInjectorTrait;
 
 class CreateAddSourcesCompileExecuteTest extends AbstractEndToEndTest
