+++ tests/Services/InvokableFactory/TestMutatorFactory.php
@@ -4,25 +4,27 @@ declare(strict_types=1);
 
 namespace App\Tests\Services\InvokableFactory;
 
-use App\Entity\Test;
-use App\Services\TestStore;
 use App\Tests\Model\EndToEndJob\Invokable;
 use App\Tests\Model\EndToEndJob\InvokableInterface;
 use App\Tests\Model\EndToEndJob\ServiceReference;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Services\EntityPersister;
 
 class TestMutatorFactory
 {
     public static function create(Test $test, callable $mutator): InvokableInterface
     {
         return new Invokable(
-            function (Test $test, TestStore $testStore, callable $mutator): Test {
+            function (Test $test, EntityPersister $entityPersister, callable $mutator): Test {
                 $test = $mutator($test);
 
-                return $testStore->store($test);
+                $entityPersister->persist($test);
+
+                return $test;
             },
             [
                 $test,
-                new ServiceReference(TestStore::class),
+                new ServiceReference(EntityPersister::class),
                 $mutator,
             ]
         );
