+++ tests/Services/TestTestRepository.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Tests\Services;
 
-use App\Entity\Test;
-use App\Repository\TestRepository;
+use webignition\BasilWorker\PersistenceBundle\Entity\Test;
+use webignition\BasilWorker\PersistenceBundle\Services\Repository\TestRepository;
 
 class TestTestRepository extends TestRepository
 {
